requirements.txt
streamlit==1.40.2
qrcode==8.0
pillow==11.0.0
